docs(wiki): add AI agent Linux & BSD component development guide
- Added `AI_AGENT_SIGMAOS_COMPONENT_DEVELOPMENT_GUIDE.md` in `WIKI/`
  outlining architectural standards (`#![no_std]`), Linux and BSD inspiration
  bridges, component workflows, and verification procedures for AI agents.
- Synchronized documentation across `WIKI/`, `wiki/`, and `wiki_repo/`.

+++ src/distro/sovereign_2040_distro_supremacy_engine.rs
@@ -185,7 +185,7 @@ impl SovereignLinux80BcachefsOpticalMeshEngine {
 
 impl Default for SovereignLinux80BcachefsOpticalMeshEngine {
     fn default() -> Self {
-        Self::new((256u64 * 1024 * 1024 * 1024) as usize as u64)
+        Self::new(256u64 * 1024 * 1024 * 1024)
     }
 }
